Add Custom Exceptions

+++ src/Bookstore.Application/Services/Implementations/DistributorService.cs
@@ -1,4 +1,6 @@
-﻿using Bookstore.DataAccess.Entities;
+﻿using Bookstore.Application.Exceptions;
+using Bookstore.DataAccess.Entities;
+using Bookstore.DataAccess.Exceptions;
 using Bookstore.DataAccess.Repositories;
 using Bookstore.DataAccess.Repositories.Implementations;
 
@@ -25,16 +27,28 @@ namespace Bookstore.Application.Services.Implementations
 
         public async Task<Distributor> CreateDistributorAsync(Distributor distributor)
         {
+            var searchResult = await _distributorRepository.GetDistributorByIdAsync(distributor.Id);
+            if (searchResult != null)
+                throw new DistributorAlreadyExistsException("A distributor with this ID already exists.");
+
             return await _distributorRepository.CreateDistributorAsync(distributor);
         }
 
         public async Task<Distributor> UpdateDistributorAsync(Distributor distributor)
         {
+            var searchResult = await _distributorRepository.GetDistributorByIdAsync(distributor.Id);
+            if (searchResult == null)
+                throw new DistributorAlreadyExistsException("This distributor does not exist.");
+
             return await _distributorRepository.UpdateDistributorAsync(distributor);
         }
 
         public async Task DeleteDistributorByIdAsync(Guid id)
         {
+            var searchResult = await _distributorRepository.GetDistributorByIdAsync(id);
+            if (searchResult == null)
+                throw new DistributorAlreadyExistsException("This distributor does not exist.");
+
             await _distributorRepository.DeleteDistributorByIdAsync(id);
         }
     }
